refactor(sync): split NoteFolderMirror along its write/import/lookup/reconcile seams
